requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
requests==2.32.5
python-dotenv==1.1.1
openai==2.1.0
loguru==0.7.3
pydantic-settings==2.10.1
transformers==4.57.0
torch==2.8.0
kernels==0.10.2
accelerate==1.10.1
hf_xet==1.1.10
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.7
sqlmodel==0.0.27
prisma==0.15.0